requirements.txt
beautifulsoup4==4.12.2
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.26
pandas==2.0.3
lxml==4.9.3
huggingface_hub>=0.24.0
scikit-learn>=1.4.0
